fix: Adjust wasm host builder location
