--- src/Uno.UI/Hosting/WebAssembly/WebAssemblyHostBuilder.wasm.cs
@@ -1,19 +1,0 @@
-﻿#nullable enable
-
-using System;
-using System.Text.RegularExpressions;
-using Uno.UI.Hosting.WebAssembly;
-
-namespace Uno.UI.Hosting;
-
-internal partial class WebAssemblyHostBuilder : IPlatformHostBuilder
-{
-	public WebAssemblyHostBuilder()
-	{
-	}
-
-	public bool IsSupported => true;
-
-	public UnoPlatformHost Create(Func<Microsoft.UI.Xaml.Application> appBuilder, Type appType)
-		=> new WebAssemblyBrowserHost(appBuilder);
-}
